requirements.txt
discord.py>=2.3.0
python-dotenv
yt-dlp
PyNaCl
ffmpeg-python
ytmusicapi
aiohttp
groq>=0.4.0